Fixed incorrect start time entry (and handling)
Out of habit, since time stamps are displayed this way in most
players, I added a leading zero to one of the segments' start time's
seconds value. This caused it to get parsed as a string instead of as
an integer.

I fixed the incorrect entry, but I also ensured the script would try
to parse these values as numbers (floats for safety's sake (and in
case someone wants to try adding fractional seconds in the future)).